# release-21.1: opt: speed up new init pattern